refactor: Added gradle.kts, changed package, new repos.
