--- src/main/kotlin/com/learnspigot/event/game/lobby/LobbyGame.kt
@@ -1,10 +1,0 @@
-package com.learnspigot.event.game.lobby
-
-import com.learnspigot.event.game.Game
-import org.bukkit.entity.Player
-
-abstract class LobbyGame : Game() {
-
-    abstract val player: Player
-
-}
