--- src/main/kotlin/com/learnspigot/event/game/main/MainGameState.kt
@@ -1,8 +1,0 @@
-package com.learnspigot.event.game.main
-
-enum class MainGameState {
-
-    COUNTDOWN,
-    ACTIVE
-
-}
