--- src/main/kotlin/com/learnspigot/event/util/Message.kt
@@ -1,5 +1,0 @@
-package com.learnspigot.event.util
-
-import net.kyori.adventure.text.format.TextColor
-
-val CHRISTMAS_RED = TextColor.color(0xF85555)
