--- src/main/kotlin/com/learnspigot/event/util/Skin.kt
@@ -1,3 +1,0 @@
-package com.learnspigot.event.util
-
-data class Skin(val texture: String, val signature: String)
