+++ src/main/kotlin/gg/flyte/event/game/lobby/LobbyGame.kt
@@ -1,0 +1,10 @@
+package gg.flyte.event.game.lobby
+
+import com.learnspigot.event.game.Game
+import org.bukkit.entity.Player
+
+abstract class LobbyGame : Game() {
+
+    abstract val player: Player
+
+}
