+++ src/main/kotlin/gg/flyte/event/game/main/MainGame.kt
@@ -1,0 +1,15 @@
+package gg.flyte.event.game.main
+
+import com.learnspigot.event.game.Game
+import org.bukkit.entity.Player
+import java.util.*
+
+abstract class MainGame : Game() {
+
+    var state = MainGameState.COUNTDOWN
+
+    val points = mutableMapOf<UUID, Int>()
+
+    abstract fun onPlayerJoin(player: Player)
+
+}
