+++ src/main/kotlin/gg/flyte/event/game/main/MainGameState.kt
@@ -1,0 +1,8 @@
+package gg.flyte.event.game.main
+
+enum class MainGameState {
+
+    COUNTDOWN,
+    ACTIVE
+
+}
