+++ src/main/kotlin/gg/flyte/event/util/Message.kt
@@ -1,0 +1,5 @@
+package gg.flyte.event.util
+
+import net.kyori.adventure.text.format.TextColor
+
+val CHRISTMAS_RED = TextColor.color(0xF85555)
