+++ src/main/kotlin/gg/flyte/event/util/Skin.kt
@@ -1,0 +1,3 @@
+package gg.flyte.event.util
+
+data class Skin(val texture: String, val signature: String)
