+++ src/main/kotlin/gg/flyte/event/util/npc/NPCClickEvent.kt
@@ -1,0 +1,22 @@
+package gg.flyte.event.util.npc
+
+import org.bukkit.entity.Player
+import org.bukkit.event.Event
+import org.bukkit.event.HandlerList
+
+class NPCClickEvent(val player: Player, val npc: NPC) : Event() {
+
+    override fun getHandlers(): HandlerList {
+        return HANDLERS
+    }
+
+    companion object {
+        val HANDLERS = HandlerList()
+
+        @JvmStatic
+        fun getHandlerList(): HandlerList {
+            return HANDLERS
+        }
+    }
+
+}
