+++ src/main/resources/plugin.yml
@@ -1,5 +1,5 @@
 name: ChristmasEvent
-version: 1.0
+version: 1.0.0
 main: com.learnspigot.event.ChristmasEvent
 api-version: 1.20
 contributors: [Stephen, Josh]
